backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
pydantic==2.11.2
pydantic-settings==2.8.1
numpy==2.3.3
faiss-cpu==1.12.0
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.20
